+++ docs/history.rst
@@ -11,6 +11,7 @@ Changed
 * Removed ``ArrayMetadataLike``, use the strict ``ArrayMetadata`` instead. (PR_20_)
 * Removed ``None`` variant for ``compiler_options`` and ``constant_arrays``, use an empty list instead. (PR_20_)
 * Module discovery is now fully automatic and does not need ``__process_modules__()`` to be defined in user classes. (PR_21_)
+* ``Snippet.from_string()`` now takes an iterable of argument names. (PR_23_)
 
 
 Added